FPLA-2056: Code review fixes
