+++ service/src/test/java/uk/gov/hmcts/reform/fpl/model/common/DocumentReferenceTest.java
@@ -15,4 +15,31 @@ class DocumentReferenceTest {
             .filename(document().originalDocumentName)
             .build());
     }
+
+    @Test
+    void shouldReturnTrueIfDocumentReferenceIsEmpty() {
+        DocumentReference documentReference = DocumentReference.builder().build();
+        assertThat(documentReference.isEmpty()).isTrue();
+    }
+
+    @Test
+    void shouldReturnFalseIfDocumentReferenceIsFullyPopulated() {
+        DocumentReference documentReference = DocumentReference.builder()
+            .filename("file.pdf")
+            .url("http://test")
+            .binaryUrl("http://test")
+            .build();
+
+        assertThat(documentReference.isEmpty()).isFalse();
+    }
+
+    @Test
+    void shouldReturnFalseIfDocumentReferenceIsPartiallyEmpty() {
+        DocumentReference documentReference = DocumentReference.builder()
+            .filename("file.pdf")
+            .binaryUrl("http://test")
+            .build();
+
+        assertThat(documentReference.isEmpty()).isFalse();
+    }
 }
